change the offset calc

+++ aio/main.c
@@ -80,7 +80,7 @@ int main(int argc, char ** argv)
     srand(5);
     clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &time1);
     for ( i = 0 ; i < REPTIMES ; i++ ) {
-        aio.aio_offset = (float)len1 * rand() / RAND_MAX;
+        aio.aio_offset = len1 * (rand() / (double)RAND_MAX);
         //printf("%ld ", aio.aio_offset);
         aio_read(&aio);
     }
@@ -108,7 +108,7 @@ int main(int argc, char ** argv)
     bytes_read = 0;
     clock_gettime(CLOCK_PROCESS_CPUTIME_ID, &time1);
     for ( i = 0 ; i < REPTIMES ; i++ ) {
-        offset = (float)len2 * rand() / RAND_MAX;
+        offset = len2 * (rand() / (double)RAND_MAX);
         //printf("%ld ", offset);
         bytes_read += pread(fd2, data, READSIZE, offset);
     }
